Move files to new training submodule
