+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from nanofold import mmcif
+from nanofold.training import mmcif
 
 
 @pytest.fixture
