+++ tests/test_backbone_update.py
@@ -1,5 +1,5 @@
 import torch
-from nanofold.model.backbone_update import BackboneUpdate
+from nanofold.training.model.backbone_update import BackboneUpdate
 
 
 def test_backbone_update():
